HDDS-7661. Refactor Test classes
