+++ hadoop-hdds/server-scm/src/test/java/org/apache/hadoop/hdds/scm/container/replication/TestMisReplicationHandler.java
@@ -1,0 +1,158 @@
+package org.apache.hadoop.hdds.scm.container.replication;
+
+import org.apache.hadoop.hdds.client.ReplicationConfig;
+import org.apache.hadoop.hdds.conf.OzoneConfiguration;
+import org.apache.hadoop.hdds.protocol.DatanodeDetails;
+import org.apache.hadoop.hdds.protocol.MockDatanodeDetails;
+import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
+import org.apache.hadoop.hdds.scm.ContainerPlacementStatus;
+import org.apache.hadoop.hdds.scm.PlacementPolicy;
+import org.apache.hadoop.hdds.scm.container.ContainerInfo;
+import org.apache.hadoop.hdds.scm.container.ContainerReplica;
+import org.apache.hadoop.hdds.scm.container.MockNodeManager;
+import org.apache.hadoop.hdds.scm.net.NodeSchema;
+import org.apache.hadoop.hdds.scm.net.NodeSchemaManager;
+import org.apache.hadoop.hdds.scm.node.NodeManager;
+import org.apache.hadoop.hdds.scm.node.NodeStatus;
+import org.apache.hadoop.hdds.scm.node.states.NodeNotFoundException;
+import org.apache.hadoop.ozone.container.common.SCMTestUtils;
+import org.apache.hadoop.ozone.protocol.commands.ReplicateContainerCommand;
+import org.apache.hadoop.ozone.protocol.commands.SCMCommand;
+import org.junit.jupiter.api.Assertions;
+import org.mockito.Mockito;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.NodeOperationalState.IN_SERVICE;
+import static org.apache.hadoop.hdds.protocol.proto.StorageContainerDatanodeProtocolProtos.SCMCommandProto.Type.replicateContainerCommand;
+import static org.apache.hadoop.hdds.scm.net.NetConstants.LEAF_SCHEMA;
+import static org.apache.hadoop.hdds.scm.net.NetConstants.RACK_SCHEMA;
+import static org.apache.hadoop.hdds.scm.net.NetConstants.ROOT_SCHEMA;
+import static org.mockito.ArgumentMatchers.anyInt;
+import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.ArgumentMatchers.anyMap;
+import static org.mockito.ArgumentMatchers.eq;
+
+/**
+ * Tests the MisReplicationHandling functionalities to test implementations.
+ */
+public abstract class TestMisReplicationHandler {
+
+  private ContainerInfo container;
+  private NodeManager nodeManager;
+  private OzoneConfiguration conf;
+
+  protected void setup(ReplicationConfig repConfig) {
+    nodeManager = new MockNodeManager(true, 10) {
+      @Override
+      public NodeStatus getNodeStatus(DatanodeDetails dd) {
+        return new NodeStatus(
+                dd.getPersistedOpState(), HddsProtos.NodeState.HEALTHY, 0);
+      }
+
+    };
+    conf = SCMTestUtils.getConf();
+    container = ReplicationTestUtil
+            .createContainer(HddsProtos.LifeCycleState.CLOSED, repConfig);
+    NodeSchema[] schemas =
+            new NodeSchema[] {ROOT_SCHEMA, RACK_SCHEMA, LEAF_SCHEMA};
+    NodeSchemaManager.getInstance().init(schemas, true);
+  }
+
+  protected abstract MisReplicationHandler getMisreplicationHandler(PlacementPolicy placementPolicy, OzoneConfiguration conf, NodeManager nodeManager);
+  protected void testMisReplication(Set<ContainerReplica> availableReplicas,
+                                  List<ContainerReplicaOp> pendingOp, int maintenanceCnt,
+                                  int misreplicationCount, int expectedNumberOfNodes) throws IOException {
+    PlacementPolicy placementPolicy = Mockito.mock(PlacementPolicy.class);
+    ContainerPlacementStatus mockedContainerPlacementStatus =
+            Mockito.mock(ContainerPlacementStatus.class);
+    Mockito.when(mockedContainerPlacementStatus.isPolicySatisfied())
+            .thenReturn(false);
+    Mockito.when(placementPolicy.validateContainerPlacement(anyList(),
+            anyInt())).thenReturn(mockedContainerPlacementStatus);
+    testMisReplication(availableReplicas, placementPolicy, pendingOp,
+            maintenanceCnt, misreplicationCount, expectedNumberOfNodes);
+  }
+
+  protected void testMisReplication(Set<ContainerReplica> availableReplicas,
+                                  PlacementPolicy mockedPlacementPolicy, List<ContainerReplicaOp> pendingOp,
+                                  int maintenanceCnt, int misreplicationCount, int expectedNumberOfNodes)
+          throws IOException {
+    MisReplicationHandler misReplicationHandler =
+            getMisreplicationHandler(mockedPlacementPolicy, conf, nodeManager);
+
+    ContainerHealthResult.MisReplicatedHealthResult result =
+            Mockito.mock(ContainerHealthResult.MisReplicatedHealthResult.class);
+    Mockito.when(result.isReplicatedOkAfterPending()).thenReturn(false);
+    Mockito.when(result.getContainerInfo()).thenReturn(container);
+    Map<ContainerReplica, Boolean> sources = availableReplicas.stream()
+            .collect(Collectors.toMap(Function.identity(),
+                    r -> {
+                      if (r.getDatanodeDetails().getPersistedOpState()
+                              == IN_SERVICE) {
+                        try {
+                          return nodeManager.getNodeStatus(r.getDatanodeDetails())
+                                  .isHealthy();
+                        } catch (NodeNotFoundException e) {
+                          throw new RuntimeException(e);
+                        }
+                      }
+                      return false;
+                    }));
+
+    Set<ContainerReplica> copy = sources.entrySet().stream()
+            .filter(Map.Entry::getValue).limit(misreplicationCount)
+            .map(Map.Entry::getKey).collect(Collectors.toSet());
+    Mockito.when(mockedPlacementPolicy.replicasToCopyToFixMisreplication(
+            anyMap())).thenAnswer(invocation -> copy);
+    Set<DatanodeDetails> remainingReplicasAfterCopy =
+            availableReplicas.stream().filter(r -> !copy.contains(r))
+                    .map(ContainerReplica::getDatanodeDetails)
+                    .collect(Collectors.toSet());
+    List<DatanodeDetails> targetNodes =
+            IntStream.range(0, expectedNumberOfNodes)
+                    .mapToObj(i -> MockDatanodeDetails.randomDatanodeDetails())
+                    .collect(Collectors.toList());
+    if (expectedNumberOfNodes > 0) {
+      Mockito.when(mockedPlacementPolicy.chooseDatanodes(
+                      Mockito.any(), Mockito.any(), Mockito.any(),
+                      eq(copy.size()), Mockito.anyLong(), Mockito.anyLong()))
+              .thenAnswer(invocation -> {
+                List<DatanodeDetails> datanodeDetails =
+                        invocation.getArgument(0);
+                Assertions.assertTrue(remainingReplicasAfterCopy
+                        .containsAll(datanodeDetails));
+                return targetNodes;
+              });
+    }
+    Map<DatanodeDetails, Integer> copyReplicaIdxMap = copy.stream()
+            .collect(Collectors.toMap(ContainerReplica::getDatanodeDetails,
+                    ContainerReplica::getReplicaIndex));
+    Map<DatanodeDetails, SCMCommand<?>> datanodeDetailsSCMCommandMap = misReplicationHandler
+            .processAndCreateCommands(availableReplicas, pendingOp,
+                    result, maintenanceCnt);
+    Assertions.assertEquals(expectedNumberOfNodes,
+            datanodeDetailsSCMCommandMap.size());
+    Assertions.assertTrue(datanodeDetailsSCMCommandMap.keySet()
+            .containsAll(targetNodes));
+    for (SCMCommand<?> command : datanodeDetailsSCMCommandMap.values()) {
+      Assertions.assertTrue(command.getType() == replicateContainerCommand);
+      ReplicateContainerCommand replicateContainerCommand =
+              (ReplicateContainerCommand) command;
+      Assertions.assertEquals(replicateContainerCommand.getContainerID(),
+              container.getContainerID());
+      DatanodeDetails replicateSrcDn =
+              replicateContainerCommand.getSourceDatanodes().stream()
+                      .findFirst().get();
+      Assertions.assertTrue(copyReplicaIdxMap.containsKey(replicateSrcDn));
+      Assertions.assertEquals(copyReplicaIdxMap.get(replicateSrcDn),
+              replicateContainerCommand.getReplicaIndex());
+    }
+  }
+}
